[core] reorganize into new timer mod
