+++ crates/mist-core/src/lib.rs
@@ -3,9 +3,7 @@ pub mod config;
 #[cfg(feature = "dialogs")]
 pub mod dialogs;
 pub mod parse;
-mod run;
-pub mod timing;
-pub use run::Run;
-mod instant;
-pub use instant::MistInstant;
-pub mod state;
+mod timer;
+pub use timer::timing;
+pub use timer::instant::MistInstant;
+pub use timer::run::Run;
